House Layout Generation using prompts/requirements.txt
streamlit
PIL
google-generativeai
python-dotenv
matplotlib
re